+++ src/main/kotlin/oggvik/mods/configurablezombieai/ConfigurableZombieAI.kt
@@ -15,6 +15,7 @@ import oggvik.mods.configurablezombieai.config.ZombieAiSavedData
 import oggvik.mods.configurablezombieai.runtime.ZombieAiRuntime
 import oggvik.mods.configurablezombieai.runtime.abnormals.targetacquisition.AbnormalTargetAcquisitionRegistry
 import oggvik.mods.configurablezombieai.runtime.abnormals.targetacquisition.behaviors.furthest.FurthestTargetAcquisitionBehavior
+import oggvik.mods.configurablezombieai.runtime.abnormals.targetacquisition.behaviors.furthestnoswitching.FurthestNoSwitchingTargetAcquisitionBehavior
 import org.apache.logging.log4j.LogManager
 import org.apache.logging.log4j.Logger
 import org.apache.commons.lang3.tuple.Pair
@@ -45,6 +46,7 @@ object ConfigurableZombieAI {
         }
 
         AbnormalTargetAcquisitionRegistry.register(FurthestTargetAcquisitionBehavior)
+        AbnormalTargetAcquisitionRegistry.register(FurthestNoSwitchingTargetAcquisitionBehavior)
 
         // Commands, zombie setup, and cache cleanup are all server-side concerns.
         FORGE_BUS.addListener(::onRegisterCommands)